CFileReading.c
Write a C program to read a file name and a number say N on command line.
if the number is negative it should print first N no. of lines from the file.
if N is positive the it should skip first N lines and print all the remaining file.

eg. If N is -5, it should print first 5 lines. If N is 5 then it should print all line starting from 6th line.